requirements.txt
# langchain==0.3.26                         
# flask==3.1.1
# sentence-transformers==4.1.0
# pypdf==5.6.1
# python-dotenv==1.1.0
# langchain-pinecone==0.2.8
# langchain-openai==0.3.24
# langchain-community==0.3.26
# -e .

langchain==1.2.15
langchain-community==0.4.1
langchain-huggingface==1.2.1
langchain-pinecone==0.2.13
langchain-classic==1.0.3
Flask==3.1.3
python-dotenv==1.2.2
sentence-transformers==5.3.0
pypdf==6.9.2
pinecone==7.3.0

